etc/requirements.txt
numpy
pandas
streamlit
openai
anthropic
python-dotenv
torch
matplotlib
plotly
langchain
langchain-openai
langgraph